chore(tga): split migrations.rs and main.rs under 500-line cap + cargo fmt (PR #613 CI fixes)
- cargo fmt -p tga: resolves 11 formatting diffs in migrations.rs,
  analyze.rs, and collect.rs.
- migrations.rs (688 lines) → migrations/ directory: migration dispatcher
  stays in migrations/mod.rs (442 lines); v17 apply_v17_pushdown_445
  function + its two regression tests moved to migrations/v17.rs (270 lines).
  All tests preserved; include_str! paths updated to ../sql/....
- main.rs (751 lines) → arg structs extracted to commands/args.rs (430
  lines): AnalyzeArgs, CollectArgs, ClassifyArgs, ReportArgs,
  DeploymentsSubcommandArgs, DeploymentsSubcommand, IncidentsSubcommandArgs,
  IncidentsSubcommand moved verbatim. main.rs reduced to 338 lines.
  Import sites in analyze.rs, classify.rs, collect.rs, report.rs updated
  from `crate::FooArgs` to `crate::commands::args::FooArgs`.
- All 111 tests + integration + doctests pass.
